requirements.txt
django
djangorestframework
streamlit
langgraph
langchain
openai
psycopg2-binary
requests